pyproject.toml
[project]
name = "movies-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Igor Ponomarev",email = "9129910001@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi (>=0.115.13,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "watchdog (>=6.0.0,<7.0.0)",
    "redis (>=6.2.0,<7.0.0)",
    "fastapi-cache (>=0.1.0,<0.2.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "aiofiles (>=24.1.0,<25.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "sqlalchemy (>=2.0.47,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
